massive update to recipes for newer conda-build.
There shall not be Python in the build env unless needed;
If Python is in the build env then $PYTHON shall be used.

also, statically linked library has to go to the build env, otherwise
compiler can't find them. Unless one install the compilers to the
host env only.

So it appears that we really shall go back to use host env only?