Use mpich on osx; abandon openmpi
The reason is openmpi produces compilers that only produces
workable binaries in conda-build, but not workable in general.

openmpi (the version we use at least)
produces binaries has an embedded rpath that needs
to be fixed by install_name_tool. Otherwise we see this error.

dyld: Library not loaded: @rpath/libmpi.40.dylib
  Referenced from: /Users/jianiniu/source/conda-channel-bccp/platform/openmpi/tests/./a.out
  Reason: image not found
Trace/BPT trap: 5

conda-build does this for us automatically, thus conda packages
works as expected. But we want to compile programs with mpicc in general.

The mpich compiler produces a good binary that works, in contrast.
It is probably easier to support only one mpi backend on the entire channel
anyways.