bump mpi4py's build number to override conda-forge.
otherwise on nersc systems we get a mpi4py from
conda-forge even if we claim our release is identical to theirs.

We probably need to use package features or something alike to
permanently fix this. But this will work as long as we keep
on track with mpi4py's releases.